### 2017-07-27 a9ad720
analysis settings bugfix
