### 2017-07-27 #335 @katyhuff
@amoeba would you be willing to review this submission? You did a great job with the Brigthway2 paper! I hope you'll let me know if you can take this on.

Alternatively, @cmutel, you might also be a great reviewer for this submission.  I recognize that, given the relationship of this submission to Brightway2, I can imagine you may have a conflict of interest. However, if you feel you can give an unconflicted review, I think the submission would benefit from someone with a strong knowledge of this software. Please let me know if you could give an unconflicted review and would be willing to do so.
