### 2017-07-27 #335 @amoeba
@katyhuff Sure. I can look at this this weekend or earlier. I would love it if @cmutel could at least comment on the correctness of this software, which I can't do confidently due to my lack of experience with LCA. I'm happy to be the reviewer though.
