### 2017-07-28 #339 @whedon
[REVIEW]: Lcopt - An interactive tool for creating fully parameterised Life Cycle Assessment (LCA) foreground models
**Submitting author:** @pjamesjoyce (<a href="http://orcid.org/0000-0002-9560-3132">P. James Joyce</a>)
**Repository:** <a href="https://github.com/pjamesjoyce/lcopt" target ="_blank">https://github.com/pjamesjoyce/lcopt</a>
**Version:** v0.4.2
