### 2017-07-29 #339 @cmutel
I can participate in this review, but currently on vacation without a computer or reliable internet until August 8th.
