### 2017-07-30 #339 @amoeba
Thanks @cmutel. I'm going through the basics right now and would really appreciate it if you could check whatever you have time for but at least the functionality item of the review checklist. I should be able to do the rest.
