### 2017-07-30 c627640
Fix lcopt_bw2_forwast_setup typo in docs
