### 2017-07-30 pjamesjoyce/lcopt#6 @pjamesjoyce
Hi @amoeba , you're quite right. It was a typo in the docs, apologies.

It's fixed in the latest commit (c627640edb8e999acf85b3ed0ba73506faeeb705)

` lcopt_bw2_forwast_setup()` is the correct version
