### 2017-07-30 pjamesjoyce/lcopt#6 @amoeba
Thanks!
