### 2017-08-04 pjamesjoyce/lcopt#7 @amoeba
README missing installation, example usage, and contributing sections
This Issue relates to the corresponding JOSS review for this software: https://github.com/openjournals/joss-reviews/issues/339

The JOSS guidelines require a high-level overview of the statement of need, installation, and example usage. In addition, guidelines for contributors are also required, though the guidelines don't specify where (I suggest the README).

I realize that you've written much of this information in your [linked documentation](http://lcopt.readthedocs.io/en/latest/)  (which is great btw) but, to meet the JOSS guidelines, could you in-line some of this information? The README.md in the root of the repo already meets the statement of need but otherwise could use the other sections.

Thanks
