### 2017-08-07 4a357b7
Update README.md to include installation, basic usage and contribution
