+++ README.md
@@ -9,3 +9,149 @@ Lcopt is a python module for creating fully parameterised LCA foreground models 
 It integrates with SimaPro and BrightWay2
 
 Online documentation, including full installation instructions, is available [here](http://lcopt.readthedocs.io)
+
+## Installation
+
+### Basic Installation
+
+For lcopt to work you should have the latest version of [brightway2](https://brightwaylca.org/) by Chris Mutel installed.
+This will make sure most of lcopts dependencies are satisfied.
+
+The instructions for installing brightway below are current as of April 2017, but check [here](https://docs.brightwaylca.org/installation.html) for the latest ones.
+
+On the command line/console, create a new environment called lcopt:
+
+```
+conda create -n lcopt python=3.6
+```
+
+
+Then activate the lcopt environment using one of these:
+```
+# Mac/Linux
+source activate lcopt
+# Windows
+activate lcopt
+```
+
+Then install brightway2:
+```
+conda install -y -q -c conda-forge -c cmutel -c haasad brightway2 jupyter
+```
+
+On windows there's an extra dependency:
+```
+conda install -y -q pywin32
+```
+
+Once brightway2 is ready to go, theres two more steps before installing lcopt itself...
+
+Install pandas:
+```
+conda install -y -q pandas
+```
+
+Update werkzeug (this is a python 3.6 thing...):
+```
+pip install -U werkzeug
+```
+
+Finally, install lcopt via pip::
+```
+pip install lcopt
+```
+
+### Linking lcopt to brightway
+
+To analyse any of the models you create in lcopt in brightway, there's an extra installation step to set up the default project and databases.
+
+Full details of this step are in the [documentation](https://lcopt.readthedocs.io/en/latest/1_installation.html#setting-up-brightway2-for-lcopt)
+
+Lcopt can create models using external LCI data from the [ecoinvent 3.3 cutoff database](http://www.ecoinvent.org/database/ecoinvent-33/ecoinvent-33.html) (ecoinvent license required) or the [FORWAST database](http://forwast.brgm.fr/)
+
+Briefly, to set up lcopt to use ecoinvent 3.3:
+
+Log into [ecoinvent.org](http://www.ecoinvent.org/login-databases.html) and go to the Files tab
+
+Download the file called `ecoinvent 3.3_cutoff_ecoSpold02.7z`
+
+Extract the file somewhere sensible on your machine, you might need to download [7-zip](http://www.7-zip.org/download.html) to extract the files.
+
+Make a note of the folder path that contains the .ecospold files, its probably `<path/extracted/to>/datasets/`
+
+Open a python console or jupyter notebook and use the setup utility function below:
+
+```python
+from lcopt.utils import lcopt_bw2_setup
+ecospold_path = r'path/to/ecospold/files' # put your own path in here
+lcopt_bw2_setup(ecospold_path)
+```
+
+To set up lcopt to use FORWAST there's no download step (the utility function downloads the latest version of the database). Simply use:
+
+```python
+from lcopt.utils import lcopt_bw2_forwast_setup
+lcopt_bw2_forwast_setup()
+```
+
+## Example Usage
+
+Below are the basic commands to get lcopt's interactive GUI up and running to create your first model. More detailed instructions are available in the [online documentation](https://lcopt.readthedocs.io/en/latest/2_use.html), including a [video runthrough](https://lcopt.readthedocs.io/en/latest/3_video_runthrough.html) of creating a simple model using the ecoinvent 3.3 database.
+
+Lcopt saves models in your current working directory, so before launching it, `cd` to the folder you want to save your models in.
+
+Lcopt is written in Python, so to use it open up a jupyter notebook or python shell and use the following commands
+
+### Importing Lcopt
+
+To import lcopt use 
+
+```python
+from lcopt import *
+```
+
+### Creating a new model
+
[…]
