### 2017-08-07 pjamesjoyce/lcopt#7 @pjamesjoyce
Sure no problem - I've added the corresponding sections to README.md in the latest commit (4a357b7 ).

I'll leave this issue open for the time being in case you think I need to add more detail.

Cheers
