### 2017-08-10 pjamesjoyce/lcopt#7 @amoeba
LGTM. Thanks!
