### 2017-08-10 #339 @amoeba
Hey @pjamesjoyce this looks great now. Thanks for responding to the two issues I raised on the repository. I've checked off everything in the reviewer checklist, including the functionality item. If @cmutel could take a second look, I would very much appreciate the help with that item. Otherwise, everything looks great.

Kudos for the extensive documentation and creating a very nice interactive tool. It looks like it represents a great deal of work and I had no troubles using it.

@katyhuff I give this submission an Accept. With respect to functionality, I've verified the software does what it says it does to the best of my knowledge but I accept that my knowledge of LCA is limited. If you'd prefer to get a topical expert review from @cmutel I'd gladly change withhold my Accept until that's been done. What do you think?
