### 2017-08-10 #339 @katyhuff
Thanks @amoeba for your review efforts and @pjamesjoyce for your prompt cooperation! I would love it if @cmutel could give this a final expert glance. I think has just gotten back into a space with internet in the last few days.
