### 2017-08-18 078c52c
fix readthedocs pdf build error
