+++ docs/index.rst
@@ -8,7 +8,7 @@ Overview
 ========
 
 lcopt - An interactive tool for creating fully parameterised Life Cycle Assessment (LCA) foreground models
---------------------------------------------------------
+----------------------------------------------------------------------------------------------------------
 
 .. image:: https://travis-ci.org/pjamesjoyce/lcopt.svg?branch=master
 	:target: https://travis-ci.org/pjamesjoyce/lcopt
