### 2017-08-18 509c6ce
add rst2pdf to rtd-requirements
