+++ rtd-requirements.txt
@@ -7,3 +7,4 @@ jinja2
 brightway2
 bw2analyzer
 eight
+rst2pdf
