### 2017-08-18 4c9087a
remove rst2pdf for readthedocs
