### 2017-08-18 pjamesjoyce/lcopt#10 @cmutel
Small improvements to docs
IMHO, you could:
- Slow down the video a bit when doing data entry, as it is hard for us old folk to see
- Put the video in the section on "Using LCOPT", as a concrete example (this section doesn't have much on the actual use of the software so far).
