### 2017-08-18 #339 @cmutel
@amoeba @katyhuff I have filed some issues on the project page; I am also getting some errors when running stuff, but still need to figure out if that is a misconfiguration on my end.
