### 2017-08-18 6202cf8
add command line entry points for lcopt_launcher (issue #8) and bw2 setup utilities
