+++ lcopt/bin/lcopt_bw2_setup_forwast.py
@@ -1,0 +1,7 @@
+from lcopt.utils import lcopt_bw2_forwast_setup
+
+def main():
+    lcopt_bw2_forwast_setup()
+
+if __name__ == "__main__":
+    main()
