### 2017-08-21 #339 @katyhuff
Thanks @cmutel. When those have been resolved, can you and @pjamesjoyce update us here?
