### 2017-08-22 cc5aa09
Add client side validation to avoid null entries (Issue #12)
