+++ lcopt/static/js/table-edit.js
@@ -1,142 +1,142 @@
 $(document).ready(function(){
 
 
-	var $TABLE = $('#table');
-	var $BTN = $('#export-btn');
-	var $EXPORT = $('#export');
+    var $TABLE = $('#table');
+    var $BTN = $('#export-btn');
+    var $EXPORT = $('#export');
 
-	// A few jQuery helpers for exporting only
-	jQuery.fn.pop = [].pop;
-	jQuery.fn.shift = [].shift;
+    // A few jQuery helpers for exporting only
+    jQuery.fn.pop = [].pop;
+    jQuery.fn.shift = [].shift;
 
-	// Post to the provided URL with the specified parameters.
-	function post(path, parameters) {
-	    var form = $('<form></form>');
+    // Post to the provided URL with the specified parameters.
+    function post(path, parameters) {
+        var form = $('<form></form>');
 
-	    form.attr("method", "post");
-	    form.attr("action", path);
+        form.attr("method", "post");
+        form.attr("action", path);
 
-	    $.each(parameters, function(key, value) {
-	        var field = $('<input></input>');
+        $.each(parameters, function(key, value) {
+            var field = $('<input></input>');
 
-	        field.attr("type", "hidden");
-	        field.attr("name", key);
-	        field.attr("value", value);
+            field.attr("type", "hidden");
+            field.attr("name", key);
+            field.attr("value", value);
 
-	        form.append(field);
-	    });
+            form.append(field);
+        });
 
-	    // The form needs to be a part of the document in
-	    // order for us to be able to submit it.
-	    $(document.body).append(form);
-	    form.submit();
-	}
+        // The form needs to be a part of the document in
+        // order for us to be able to submit it.
+        $(document.body).append(form);
+        form.submit();
+    }
 
-	//$BTN.click(function () {
+    //$BTN.click(function () {
 
-	function save_parameters(){
-	  var $rows = $TABLE.find('tr');
-	  var headers = [];
-	  var data = [];
-	  
-	  // Get the headers (add special header logic here)
-	  $($rows.shift()).find('th').each(function () {
-	    headers.push($(this).text());
-	  });
-	  
-	  console.log($rows)
+    function save_parameters(){
+      var $rows = $TABLE.find('tr');
+      var headers = [];
+      var data = [];
+      
+      // Get the headers (add special header logic here)
+      $($rows.shift()).find('th').each(function () {
+        headers.push($(this).text());
+      });
+      
+      console.log($rows)
 
-	  // Turn all existing rows into a loopable array
-	  $rows.each(function () {
-	  	var $td, h;
-	  	if ($(this).hasClass('data_row')){
-		    $td = $(this).find('.export');
-		    h = {};
-		    
-		    // Use the headers from earlier to name our hash keys
-		    headers.forEach(function (header, i) {
-		      h[header] = $td.eq(i).text();   
-		    });
-		    
-		    data.push(h);
+      // Turn all existing rows into a loopable array
+      $rows.each(function () {
+        var $td, h;
+        if ($(this).hasClass('data_row')){
+            $td = $(this).find('.export');
+            h = {};
+            
+            // Use the headers from earlier to name our hash keys
+            headers.forEach(function (header, i) {
+              h[header] = $td.eq(i).text();   
+            });
+            
+            data.push(h);
 
-		}else if($(this).hasClass('function_row')){
-			$td = $(this).find('.export');
-		    h = {};
-		    
-		    // Use the headers from earlier to name our hash keys
-		    headers.forEach(function (header, i) {
-		      if(i<=2){
-		        h[header] = $td.eq(i).text();  
-		  	  }else{
-		  	  	h[header] = "[FUNCTION]"; //$td.eq(i).text();  	
[…]
