### 2017-08-22 02c6cdb
add example model, and ability to open it from lcopt-launcher (Issue #11)
