+++ lcopt/bin/lcopt_launcher.py
@@ -1,74 +1,106 @@
 from tkinter import *
 from tkinter import ttk
 from tkinter import filedialog, simpledialog
+from lcopt.utils import DEFAULT_DB_NAME, FORWAST_DB_NAME
+from lcopt import LcoptModel
+from brightway2 import *
+import os
+from pathlib import Path
+
+
+def check_databases():
+
+    return DEFAULT_DB_NAME in projects, FORWAST_DB_NAME in projects
 
 
 def main():
-	
-	def create_model(*args):
-		print("create")
-		root.withdraw()
-		model_name = simpledialog.askstring("New", "Enter a name for your model")
-		if model_name:
-			from lcopt import LcoptModel
-			model = LcoptModel(model_name)
-			model.launch_interact()
-		
-		root.destroy()
+    
+    CHECK_ECOINVENT, CHECK_FORWAST = check_databases()
 
-	def load_model(*args):
-		print("load")
-		root.withdraw()
+    def create_model(*args):
+        print("Create")
+        root.withdraw()
+        model_name = simpledialog.askstring("New", "Enter a name for your model")
+        if model_name:
+            model = LcoptModel(model_name)
+            model.launch_interact()
+        
+        root.destroy()
 
-		titleString = "Choose a model to open"
-		filetypesList = [('Lcopt model files', '.lcopt')]
-		file_path = filedialog.askopenfilename(title = titleString, filetypes = filetypesList )
-		print (file_path)
+    def load_model(*args):
+        print("Load")
+        root.withdraw()
 
-		if file_path:
-			from lcopt import LcoptModel
-			model = LcoptModel(load = file_path)
-			model.launch_interact()
+        titleString = "Choose a model to open"
+        filetypesList = [('Lcopt model files', '.lcopt')]
+        file_path = filedialog.askopenfilename(title=titleString, filetypes=filetypesList)
+        print (file_path)
 
-		root.destroy()
+        if file_path:
+            model = LcoptModel(load=file_path)
+            model.launch_interact()
 
+        root.destroy()
 
-	root = Tk()
-	root.title("LCOPT Launcher")
-	
-	screen_width = root.winfo_screenwidth()
-	screen_height = root.winfo_screenheight()
-	initial_width = 275
-	initial_height = 125
-	initial_x = int(screen_width/2 - initial_width/2)
-	initial_y = int(screen_height/2 - initial_height/2)
+    def load_example(*args):
+        print("Load example")
 
-	root.geometry('{}x{}+{}+{}'.format(initial_width, initial_height, initial_x, initial_y))
+        root.withdraw()
 
-	#TODO: Maybe figure out how to use a custom icon - not urgent 
-	#img = PhotoImage(file = r'C:\Users\pjjoyce\Dropbox\04. REDMUD IP LCA Project\04. Modelling\lcopt\static\img\lcoptIcon2.gif')
-	#root.tk.call('wm', 'iconphoto', root._w, img)
-	
+        asset_path = os.path.join(Path(os.path.dirname(os.path.realpath(__file__))).parent, 'assets')
 
-	mainframe = ttk.Frame(root, padding="20 20 20 20")#padding="3 3 12 12")
-	mainframe.grid(column=0, row=0, sticky=(N, W, E, S))
-	mainframe.columnconfigure(0, weight=1)
-	mainframe.rowconfigure(0, weight=1)
+        if CHECK_ECOINVENT:
+            file_path = os.path.join(asset_path, 'ecoinvent_example.lcopt')
+        elif CHECK_FORWAST:
+            file_path = os.path.join(asset_path, 'forwast_example.lcopt')
+        else:
[…]
