+++ lcopt_bw2_setup_forwast.py
@@ -1,4 +1,0 @@
-from lcopt.utils import lcopt_bw2_forwast_setup
-
-if __name__ == "__main__":
-    lcopt_bw2_forwast_setup()
