### 2017-08-22 cbe5da8
Update docs to reflect changes and solve Issue #10
