+++ docs/2_Getting_Started.rst
@@ -1,0 +1,42 @@
+.. _getting_started:
+
+===============
+Getting Started
+===============
+
+.. sidebar:: Set up the background databases
+
+	:ref:`Set up with ecoinvent<bw2_setup_step>`
+	
+	:ref:`Set up with FORWAST<forwast_setup_step>`
+
+The best way to get started with Lcopt is to explore one of the example models.
+
+First make sure you've set up either ecoinvent or FORWAST as a background database.
+
+.. highlight:: console
+
+Then activate your lcopt environment, and use the command line command `lcopt-launcher` to launch Lcopt ::
+
+	activate lcopt
+	cd C:\Users\pjjoyce\Documents\01_Lcopt_models
+	lcopt-launcher
+
+This dialog will appear:
+
+.. image:: assets/lcopt_launcher.jpg
+
+Choose *Open Example Model*
+
+The video below runs through the steps required to make the example model.
+
+.. _video:
+
+Video
+-----
+.. raw:: html
+
+	<video height="480" poster="_static/runthrough_title.jpg" controls>
+	<source src="_static/Lcopt_runthrough.mp4" type="video/mp4">
+	</video>
+
