+++ docs/3_video_runthrough.rst
@@ -1,11 +1,0 @@
-================
-Video runthrough
-================
-
-.. raw:: html
-
-	<video height="480" poster="_static/runthrough_title.jpg" controls>
-	<source src="_static/Lcopt_runthrough2.mp4" type="video/mp4">
-	</video>
-
-
