### 2017-08-22 e732174
version 0.4.1 ready for conda packaging
