+++ setup.py
@@ -35,7 +35,7 @@ print(my_package_files)
 
 setup(
     name='lcopt',
-    version="0.4.0",
+    version="0.4.1",
     packages=packages,
     author="P. James Joyce",
     author_email="pjamesjoyce@gmail.com",
@@ -52,7 +52,7 @@ setup(
     #],
     include_package_data=True, 
     url="https://github.com/pjamesjoyce/lcopt/",
-    download_url="https://github.com/pjamesjoyce/lcopt/archive/0.1.tar.gz",
+    download_url="https://github.com/pjamesjoyce/lcopt/archive/0.4.1.tar.gz",
     long_description=open('README.md').read(),
     description='An interactive tool for creating fully parameterised Life Cycle Assessment (LCA) foreground models',
     keywords=['LCA', 'Life Cycle Assessment', 'Foreground system', 'Background system', 'Foreground model', 'Fully parameterised'],
