### 2017-08-22 c79fcbb
add bin folder to setup.py
