### 2017-08-22 a770615
add bin to manifest.in
