+++ build/lib/lcopt/interact.py
@@ -1,675 +1,0 @@
-from flask import Flask, request, render_template, redirect
-import webbrowser
-import json
-from ast import literal_eval
-from lcopt.io import exchange_factory
-from collections import OrderedDict
-from itertools import groupby
-
-from lcopt.bw2_export import Bw2Exporter
-
-class FlaskSandbox():
-    
-    def __init__(self, modelInstance):
-        
-        self.modelInstance = modelInstance
-        self.get_sandbox_variables()
-        
-        # Set up the dictionary of actions that can be processed by POST requests
-        self.postActions = {
-            'savePosition':self.savePosition,
-            'saveModel':self.saveModel,
-            'newProcess':self.newProcess,
-            'echo':self.echo,
-            'searchEcoinvent':self.searchEcoinvent,
-            'searchBiosphere':self.searchBiosphere,
-            'newConnection': self.newConnection,
-            'addInput':self.addInput,
-            'inputLookup':self.inputLookup,
-            'parse_parameters': self.parameter_parsing,
-            'create_function': self.create_function,
-            'add_parameter': self.add_parameter,
-            'simaPro_export': self.simaPro_export,
-        }
-        
-        #print (self.modelInstance.newVariable)
-        
-    def shutdown_server(self):
-        func = request.environ.get('werkzeug.server.shutdown')
-        if func is None:
-            raise RuntimeError('Not running with the Werkzeug Server')
-        func()
-        
-    def output_code(self, process_id):
-        
-        exchanges = self.modelInstance.database['items'][process_id]['exchanges']
-        
-        production_filter = lambda x:x['type'] == 'production'
-           
-        code = list(filter(production_filter, exchanges))[0]['input'][1]
-        
-        return code
-    
-    def get_sandbox_variables(self):
-        m = self.modelInstance
-        db = m.database['items']
-        matrix = m.matrix
-        
-        sandbox_positions = m.sandbox_positions
-
-
-        products = OrderedDict((k,v) for k, v in db.items() if v['type'] == 'product')
-        product_codes = [k[1] for k in products.keys()]
-
-
-        processes = OrderedDict((k,v) for k, v in db.items() if v['type'] == 'process')
-        process_codes = [k[1] for k in processes.keys()]
-        process_name_map = {k[1]:v['name'] for k, v in processes.items()}
-
-
-        #note this maps from output code to process
-        process_output_map = {self.output_code(x): x[1] for x in processes.keys()}
-        self.reverse_process_output_map = {value:key for key, value in process_output_map.items()}
-
-
-        intermediates = {k:v for k, v in products.items() if v['lcopt_type'] == 'intermediate'}
-        intermediate_codes = [k[1] for k in intermediates.keys()]
-        intermediate_map = {k[1]: v['name'] for k, v in intermediates.items()}
-
-        #process_output_name_map = {process_code: output_name for x in processes.keys()}
-        process_output_name_map = {x[1]: intermediate_map[self.reverse_process_output_map[x[1]]] for x in processes.keys()}
-
-
-        inputs = OrderedDict((k,v) for k, v in products.items() if v['lcopt_type'] == 'input')
-        input_codes = [k[1] for k in inputs.keys()]
-        input_map = {k[1]: v['name'] for k, v in inputs.items()}
-        self.reverse_input_map = {value:key for key, value in input_map.items()}
-
-
-        biosphere = OrderedDict((k,v) for k, v in products.items() if v['lcopt_type'] == 'biosphere')
-        biosphere_codes = [k[1] for k in biosphere.keys()]
-        biosphere_map = {k[1]: v['name'] for k, v in biosphere.items()}
[…]
