+++ build/lib/lcopt/static/js/status.js
@@ -1,129 +1,0 @@
-$(document).ready(function(){
-	/*
-	$.getJSON( "/status.json", function(data) {
-		console.log(data)
-		$statusbox = $('#status_box_outer');
-
-		$statusbox.append($('<div class = "status_box_item row" id = "status_has_model"></div>'))
-
-		$has_model = $statusbox.find('#status_has_model')
-
-		$has_model.append('<div class="status_title col-xs-10">Model exists</div>')
-			if(data.has_model){$has_model.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_true">check</i></div>')}
-				else{
-					$has_model.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_false">close</i></div>')
-				}
-
-		$statusbox.append($('<div class = "status_box_item row" id = "status_model_has_impacts"></div>'))
-		$model_has_impacts = $statusbox.find('#status_model_has_impacts')
-
-		$model_has_impacts.append('<div class="status_title col-xs-10">Model is linked</div>')
-			if(data.model_has_impacts){$model_has_impacts.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_true">check</i></div>')}
-				else{
-					$model_has_impacts.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_false">close</i></div>')
-				}
-
-		$statusbox.append($('<div class = "status_box_item row" id = "status_model_has_parameters"></div>'))
-		$model_has_parameters = $statusbox.find('#status_model_has_parameters')
-
-		$model_has_parameters.append('<div class="status_title col-xs-10">Model is parameterised</div>')
-			if(data.model_has_parameters){$model_has_parameters.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_true">check</i></div>')}
-				else{
-					$model_has_parameters.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_false">close</i></div>')
-				}
-
-		$statusbox.append($('<div class = "status_box_item row" id = "status_model_has_functions"></div>'))
-		$model_has_functions =  $statusbox.find('#status_model_has_functions')
-
-		$model_has_functions.append('<div class="status_title col-xs-10">Model hase functions</div>')
-			if(data.model_has_functions){$model_has_functions.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_true">check</i></div>')}
-				else{
-					$model_has_functions.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_false">close</i></div>')
-				}
-
-		$statusbox.append($('<div class = "status_box_item row" id = "status_model_is_runnable"></div>'))
-		$model_is_runnable = $statusbox.find('#status_model_is_runnable')
-
-		$model_is_runnable.append('<div class="status_title col-xs-10">Model is runnable</div>')
-			if(data.model_is_runnable){$model_is_runnable.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_true">check</i></div>')}
-				else{
-					$model_is_runnable.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_false">close</i></div>')
-				}
-
-		$statusbox.append($('<div class = "status_box_item row" id = "status_model_is_fully_formed"></div>'))
-		$model_is_fully_formed = $statusbox.find('#status_model_is_fully_formed')
-
-		$model_is_fully_formed.append('<div class="status_title col-xs-10">Model is fully formed</div>')
-			if(data.model_is_fully_formed){$model_is_fully_formed.append('<div class = "status_item col-xs-6 col-xs-2"><i class="material-icons is_true">check</i></div>')}
-				else{
[…]
