### 2017-08-22 pjamesjoyce/lcopt#10 @pjamesjoyce
Looking back at the video you're absolutely right! I've slowed it down and added some more descriptive captions relating to the data entry. It's now more of a *how-to* video - it was previously more showcasing what it could do.

I've also moved the video to the newly created [Getting Started](https://lcopt.readthedocs.io/en/latest/2_Getting_Started.html) section and made a couple of edits to the [Using Lcopt](https://lcopt.readthedocs.io/en/latest/3_use.html) section.

I'll leave this open for you to close if you're happy with the updates.
