### 2017-08-24 pjamesjoyce/lcopt#10 @pjamesjoyce
Link this issue to openjournals/joss-reviews#339
