### 2017-08-24 #339 @pjamesjoyce
I've referenced the issues to here so they're all in one place.
I've left pjamesjoyce/lcopt#10 open as it included changes to the docs so whether it's resolved is subjective rather than technical. If @cmutel and/or @amoeba could take a quick look and close the issue if it's ok that'd be great.
