### 2017-08-24 #339 @amoeba
That all looks great. Thanks for helping out @cmutel. And thanks for the quick responses to edits, @pjamesjoyce. I'm good here and I approve this submission. cc @katyhuff

And, as I said before, this is a really well put-together piece of software, especially the polished nature of the UI and the highly-detailed docs. Well done!
