### 2017-08-25 #339 @pjamesjoyce
Thanks everyone (@katyhuff, @amoeba, @cmutel)!

Here's the DOI of the Zenodo archive -  10.5281/zenodo.848529

[![DOI](https://zenodo.org/badge/76573839.svg)](https://zenodo.org/badge/latestdoi/76573839)
